✨  feat: Add `2021 Day 6`!
Found `Part #2` to be quite hard.
Spent nearly six hours trying to find a solution to no avail.
Rewrote a recursive solution that (hopefully) reduced space complexity.
More optimization, but time complexity remained largely the same.
Ran the recursive solution for 6 hours and brute-forced answer.
Finally, thanks to a tip by @ZhYGuoL, found the efficient solution.

The recursive solution is in `pt1` and final solution is in `pt2`